Add subprompt weights / attention to img2img as well
